requirements.txt
psycopg2-binary
reportlab
matplotlib
